refactor: CommonDao 인터페이스 불필요한 default 메서드 삭제
