+++ src/main/java/roomescape/dto/request/PopularThemeRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 import jakarta.validation.constraints.Max;
