+++ src/test/java/roomescape/service/fake/FakeThemeDao.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
 import roomescape.domain.vo.Name;
-import roomescape.dto.PopularThemeRequestDto;
+import roomescape.dto.request.PopularThemeRequestDto;
 import roomescape.dto.response.AvailableTimeResponseDto;
 
 public class FakeThemeDao implements ThemeDao {
